Migrate repo to typescript completely
