--- src/data/aggregated.d.ts
@@ -1,4 +1,0 @@
-declare module './aggregated' {
-    const data: any[];
-    export default data;
-}
