+++ src/processors/enrich/inferData.ts
@@ -53,7 +53,7 @@ const recipes: Recipe[] = [
       "ALP",
       "WBC",
     ],
-    (...args) => getPhenotypicAge(...args, 37).toFixed(1),
+    (a, b, c, d, e, f, g, h, i, age) => getPhenotypicAge(a, b, c, d, e, f, g, h, i, age).toFixed(1),
   ],
   [
     "PhenoAge2",
@@ -68,7 +68,7 @@ const recipes: Recipe[] = [
       "ALP",
       "WBC",
     ],
-    (...args) => getPhenotypicAge2(...args, 37).toFixed(1),
+    (a, b, c, d, e, f, g, h, i, age) => getPhenotypicAge2(a, b, c, d, e, f, g, h, i, age).toFixed(1),
   ],
   ["VLDL", ["Triglyceride"], (triglyceride) => (triglyceride / 5).toFixed(0)],
   ["Age", [], (age) => age.toFixed(1)],
