--- src/processors/enrich/phenoAge.d.ts
@@ -1,2 +1,0 @@
-export function getPhenotypicAge(...args: any[]): number;
-export function getPhenotypicAge2(...args: any[]): number;
