--- src/processors/post/index.d.ts
@@ -1,2 +1,0 @@
-const postProcess: (entries: Entry[]) => Entry[];
-export default postProcess;
