--- src/processors/post/tag.d.ts
@@ -1,3 +1,0 @@
-export const tagKeys: string[];
-export const unsortedTags: string[];
-export const tagDescription: Record<string, string>;
