--- src/processors/pre/convertName.js
@@ -1,56 +1,0 @@
-const nameMapper = {
-  "Số lượng hồng cầu (RBC)": "RBC",
-  "Gamma GT": "GGT",
-  "Glucose máu": "Glucose",
-  "Alkaline Phosphatase": "ALP",
-  "HDL-Cholesterol": "HDL",
-  "LDL-Cholesterol": "LDL",
-  "Thể tích trung bình HC (MCV)": "MCV",
-  "Lượng huyết sắc tố (Hb)": "Hb",
-  "Thể tích khối hồng cầu (HCT)": "HCT",
-  "Lượng Hb trung bình HC (MCH)": "MCH",
-  "Nồng độ Hb trung bình HC (MCHC)": "MCHC",
-  "Độ phân bố HC (RDW-CV)": "RDW-CV",
-  "Thể tích khối tiểu cầu (PCT)": "PCT",
-  "Thể tích trung bình TC (MPV)": "MPV",
-  "Độ phân bố TC (PDW)": "PDW",
-  "Số lượng bạch cầu (WBC)": "WBC",
-  "Số lượng tiểu cầu (PLT)": "PLT",
-  "Độ phân bố HC (RDW-SD)": "RDW-SD",
-  "eGFR (Mức lọc cầu thận)": "eGFR",
-  "IGF-1 miễn dịch": "IGF-1",
-  "BUN (Blood Urea Nitrogen)": "BUN",
-  "Homocysteine total": "Homocysteine",
-  "Calci máu": "Calcium",
-  "Tỷ lệ % bạch cầu trung tính": "% Neutrophil",
-  "Tỷ lệ % bạch cầu Lympho": "% Lymphocyte",
-  "Tỷ lệ % bạch cầu Mono": "% Monocyte",
-  "Tỷ lệ % bạch cầu ái toan": "% Eosinophil",
-  "Tỷ lệ % bạch cầu ái kiềm": "% Basophil",
-  "% Lympho không điển hình": "% Lympho không điển hình",
-  "Sắt huyết thanh": "Serum iron",
-  "Số lượng bạch cầu trung tính": "SL Neutrophil",
-  "Số lượng bạch cầu Lympho": "SL Lymphocyte",
-  "Số lượng bạch cầu Mono": "SL Monocyte",
-  "Số lượng bạch cầu ái toan": "SL Eosinophil",
-  "Số lượng bạch cầu ái kiềm": "SL Basophil",
-  "Protein máu": "Protein Total",
-  "Albumin máu": "Albumin",
-  "Số lượng tiểu cầu có KT lớn": "SL tiểu cầu KT lớn",
-  "Ure máu": "Ure",
-  "Creatinin máu": "Creatinin",
-  "Uric acid máu": "Uric",
-  "Natri máu": "Natri",
-  "Kali máu": "Kali",
-  "Clo máu": "Clo",
-  "Ketone/Máu": "Ketone",
-  "Định lượng Cortisol máu": "Cortisol",
-  "Khả năng gắn sắt toàn phần": "TIBC",
-  "SL tiểu cầu KT lớn": "P-LCC",
-  "Tỉ lệ tiểu cầu có KT lớn": "P-LCR",
-  "Tỉ lệ Prothrombin": "Prothrombin"
-};
-
-export default ([name, values, unit, extra]) => {
-  return [nameMapper[name] || name, values, unit, extra];
-};
