--- src/processors/pre/excludes.js
@@ -1,15 +1,0 @@
-export default [
-  // "TSH",
-  "HBsAg",
-  "SL TB non lớn",
-  "% Các TB non lớn",
-  // "LDH",
-  "HCV Ab",
-  "HBsAb định lượng",
-  "% Lympho không điển hình",
-  "SL Lympho không điển hình",
-  "BUN",
-  // "Tranferrin",
-  "T4",
-  "IGF-1",
-];
