+++ src/processors/pre/excludes.ts
@@ -1,0 +1,17 @@
+const excludes: string[] = [
+  // "TSH",
+  "HBsAg",
+  "SL TB non lớn",
+  "% Các TB non lớn",
+  // "LDH",
+  "HCV Ab",
+  "HBsAb định lượng",
+  "% Lympho không điển hình",
+  "SL Lympho không điển hình",
+  "BUN",
+  // "Tranferrin",
+  "T4",
+  "IGF-1",
+];
+
+export default excludes;
