--- src/processors/pre/index.d.ts
@@ -1,2 +1,0 @@
-const preProcess: (entries: Entry[]) => Entry[];
-export default preProcess;
